EDGETTS/requirements.txt
gradio
pydub
edge-tts
openai
python-dotenv
pandas